refactor: event tab screens
