+++ app/src/main/java/com/github/se/assocify/ui/screens/event/scheduletab/EventScheduleScreen.kt
@@ -1,4 +1,4 @@
-package com.github.se.assocify.ui.screens.event.schedule
+package com.github.se.assocify.ui.screens.event.scheduletab
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
